Added GlobalScene singleton.
Global scene is loaded by the current SceneController (e.g.
LevelController) if it hasn't been loaded yet. Once complete the Init
function is called.

Also added a roof to the external level geometry to allow the use of the
directional light.

Ensured that the player can't fire the gun whilst sprinting.